requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
python-multipart
fastapi-mail